+++ src/main/java/net/silentchaos512/gems/init/ModItems.java
@@ -34,7 +34,8 @@ public final class ModItems {
         registerGemItems(reg, Gems::getItem, Gems::getName);
         registerGemItems(reg, Gems::getShard, gem -> gem.getName() + "_shard");
 
-//        register(reg, "soul_gem", ItemSoulGem.INSTANCE.get());
+        // temp placeholder
+        register(reg, "soul_gem", new Item(new Item.Properties()));
 
         for (CraftingItems item : CraftingItems.values()) {
             register(reg, item.getName(), item.asItem());